refactor: Seclude ServerlessError into lib/serverless-error.js
