+++ lib/Serverless.js
@@ -14,7 +14,7 @@ const Utils = require('./classes/Utils');
 const Service = require('./classes/Service');
 const Variables = require('./classes/Variables');
 const ConfigSchemaHandler = require('./classes/ConfigSchemaHandler');
-const ServerlessError = require('./classes/Error').ServerlessError;
+const ServerlessError = require('./serverless-error');
 const version = require('./../package.json').version;
 const isStandaloneExecutable = require('./utils/isStandaloneExecutable');
 const resolveCliInput = require('./utils/resolveCliInput');
