+++ lib/classes/Error.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const chalk = require('chalk');
+const ServerlessError = require('../serverless-error');
 
 const consoleLog = (message) => {
   console.log(message); // eslint-disable-line no-console
@@ -22,19 +23,6 @@ const writeMessage = (title, message) => {
 
   consoleLog(' ');
 };
-
-class ServerlessError extends Error {
-  constructor(message, code) {
-    super(message);
-    this.code = code;
-    Error.captureStackTrace(this, this.constructor); // Not needed in Node.js v8+
-  }
-}
-Object.defineProperty(ServerlessError.prototype, 'name', {
-  value: ServerlessError.name,
-  configurable: true,
-  writable: true,
-});
 
 module.exports.ServerlessError = ServerlessError;
