+++ lib/loadEnv.js
@@ -2,7 +2,7 @@
 
 const dotenv = require('dotenv');
 const path = require('path');
-const ServerlessError = require('./classes/Error').ServerlessError;
+const ServerlessError = require('./serverless-error');
 
 const isMissingFileError = (error) => error.code === 'ENOENT';
