+++ lib/plugins/aws/package/compile/events/apiGateway/lib/hack/updateStage.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const BbPromise = require('bluebird');
-const ServerlessError = require('../../../../../../../../classes/Error').ServerlessError;
+const ServerlessError = require('../../../../../../../../serverless-error');
 
 const defaultApiGatewayLogFormat = [
   'requestId: $context.requestId',
