+++ lib/plugins/aws/package/compile/events/websockets/lib/stage.js
@@ -2,7 +2,7 @@
 
 const BbPromise = require('bluebird');
 const ensureApiGatewayCloudWatchRole = require('../../lib/ensureApiGatewayCloudWatchRole');
-const ServerlessError = require('../../../../../../../classes/Error').ServerlessError;
+const ServerlessError = require('../../../../../../../serverless-error');
 
 const defaultLogLevel = 'INFO';
 const validLogLevels = new Set(['INFO', 'ERROR']);
