+++ lib/plugins/aws/utils/resolveCfImportValue.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const ServerlessError = require('../../../classes/Error').ServerlessError;
+const ServerlessError = require('../../../serverless-error');
 
 function resolveCfImportValue(provider, name, sdkParams = {}) {
   return provider.request('CloudFormation', 'listExports', sdkParams).then((result) => {
