+++ lib/plugins/aws/utils/resolveCfRefValue.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const ServerlessError = require('../../../classes/Error').ServerlessError;
+const ServerlessError = require('../../../serverless-error');
 
 function resolveCfRefValue(provider, resourceLogicalId, sdkParams = {}) {
   return provider
