+++ lib/plugins/create/create.js
@@ -5,7 +5,7 @@ const path = require('path');
 const fse = require('fs-extra');
 const untildify = require('untildify');
 
-const ServerlessError = require('../../classes/Error').ServerlessError;
+const ServerlessError = require('../../serverless-error');
 const createFromTemplate = require('../../utils/createFromTemplate');
 const download = require('../../utils/downloadTemplateFromRepo');
 const renameService = require('../../utils/renameService').renameService;
