+++ lib/utils/downloadTemplateFromRepo.js
@@ -10,7 +10,7 @@ const qs = require('querystring');
 const fetch = require('node-fetch');
 const spawn = require('child-process-ext/spawn');
 const renameService = require('./renameService').renameService;
-const ServerlessError = require('../classes/Error').ServerlessError;
+const ServerlessError = require('../serverless-error');
 const copyDirContentsSync = require('./fs/copyDirContentsSync');
 const dirExistsSync = require('./fs/dirExistsSync');
 const log = require('./log/serverlessLog');
