+++ lib/utils/getServerlessConfigFile.js
@@ -8,7 +8,7 @@ const memoizee = require('memoizee');
 const spawn = require('child-process-ext/spawn');
 const fileExists = require('./fs/fileExists');
 const readFile = require('./fs/readFile');
-const ServerlessError = require('../classes/Error').ServerlessError;
+const ServerlessError = require('../serverless-error');
 
 const getConfigFilePath = async (servicePath, options = {}) => {
   if (options.config) {
