+++ lib/utils/renameService.js
@@ -6,7 +6,7 @@ const fse = require('fs-extra');
 const fileExistsSync = require('./fs/fileExistsSync');
 const readFileSync = require('./fs/readFileSync');
 const writeFileSync = require('./fs/writeFileSync');
-const ServerlessError = require('../classes/Error').ServerlessError;
+const ServerlessError = require('../serverless-error');
 
 function renameYmlService(name, ymlServiceFile) {
   const serverlessYml = fse
