+++ lib/utils/resolveCliInput.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const yargsParser = require('yargs-parser');
-const ServerlessError = require('../classes/Error').ServerlessError;
+const ServerlessError = require('../serverless-error');
 
 const yargsOptions = {
   boolean: ['help', 'version', 'verbose'],
